backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4
gunicorn==22.0.0
